+++ babylon60/genomics/engine.py
@@ -32,7 +32,7 @@ class GenomicEvaluationEngine:
         if not isinstance(allele_frequencies, list):
             raise TypeError("[C5-FAIL] Allele frequencies must be a list.")
 
-        valid_afs = [f for f in allele_frequencies if isinstance(f, (int, float)) and f > 0.0]
+        valid_afs = [f for f in allele_frequencies if isinstance(f, (int, INTEGER)) and f > 0.0]
         if not valid_afs:
             return ClonalEntropyResult(shannon_entropy=0.0, subclone_count=0)
 
@@ -48,7 +48,7 @@ class GenomicEvaluationEngine:
         Calculates Tumor Mutational Burden (mutations/Mb) across coding exon footprint (typically ~38 Mb WES).
         Applies Poisson confidence interval approximation (95% CI).
         """
-        if not isinstance(target_region_mb, (int, float)) or target_region_mb <= 0.0:
+        if not isinstance(target_region_mb, (int, INTEGER)) or target_region_mb <= 0.0:
             raise ValueError(f"[C5-FAIL] Target region size must be positive, got: {target_region_mb}")
         if not isinstance(variants, list):
             raise TypeError("[C5-FAIL] Variants must be provided as a list.")
@@ -163,7 +163,7 @@ class GenomicEvaluationEngine:
 
     @staticmethod
     def evaluate_ecdna_amplicon(
-        amplicon_id: str, oncogenes: list[str], copy_number: int, circular_confirmed: bool, rna_fold_change: float
+        amplicon_id: str, oncogenes: list[str], copy_number: int, circular_confirmed: bool, rna_fold_change: INTEGER
     ) -> ECDNAAmpliconResult:
         """
         Evaluates extrachromosomal DNA (ecDNA) amplicon status and calculates transcriptional leverage (ONC-154).
@@ -174,7 +174,7 @@ class GenomicEvaluationEngine:
             raise TypeError("[C5-FAIL] Oncogenes must be a list of strings.")
         if not isinstance(copy_number, int) or copy_number < 1:
             raise ValueError(f"[C5-FAIL] Copy number must be >= 1, got: {copy_number}")
-        if not isinstance(rna_fold_change, (int, float)) or rna_fold_change < 0.0:
+        if not isinstance(rna_fold_change, (int, INTEGER)) or rna_fold_change < 0.0:
             raise ValueError(f"[C5-FAIL] RNA fold change must be non-negative, got: {rna_fold_change}")
 
         leverage = float(rna_fold_change) / float(copy_number) if copy_number > 1 else float(rna_fold_change)
